requirements.txt

fastapi
uvicorn
pydantic
pymongo
passlib
python-dotenv
pyjwt
aiohttp 
sounddevice 
numpy
scipy
pydub
uvicorn[standard]
elevenlabs==1.3.1
# sudo apt update
# sudo apt install ffmpeg
pydantic[email]
python-multipart
openai == 0.28
aiofiles
sendgrid
aioboto3
noisereduce
motor
# annotated-types==0.6.0
# anyio==4.3.0
# asttokens==2.4.1
# boto3==1.34.84
# botocore==1.34.84
# certifi==2024.2.2
# charset-normalizer==3.3.2
# colorama==0.4.6
# decorator==5.1.1
# elevenlabs==1.0.5
# exceptiongroup==1.2.0
# executing==2.0.1
# h11==0.14.0
# httpcore==1.0.5
# httpx==0.27.0
# idna==3.7
# ipython==8.23.0
# jedi==0.19.1
# jmespath==1.0.1
# matplotlib-inline==0.1.7
# parso==0.8.4
# prompt-toolkit==3.0.43
# pure-eval==0.2.2
# pydantic==2.7.0
# pydantic_core==2.18.1
# Pygments==2.17.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.0.1
# requests==2.31.0
# s3transfer==0.10.1
# six==1.16.0
# sniffio==1.3.1
# stack-data==0.6.3
# traitlets==5.14.2
# typing_extensions==4.11.0
# urllib3==2.2.1
# wcwidth==0.2.13
# websockets==12.0
